pyproject.toml
[project]
name = "scales_driver_async"
version = "0.0.6"
authors = [
  { name="Aleksey Kravtsun", email="kravtsun.aleksey@gmail.com" },
]
dependencies = [
  "pyserial-asyncio>=0.6",
]
description = "Asynchronous scale driver"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Development Status :: 4 - Beta",
]

[project.urls]
Homepage = "https://github.com/kr-aleksey/ScalesDriverAsync"
AuthorGitHub = "https://github.com/kr-aleksey"
